// frontend/types.ts
export type MessageProps = {
  id: string;
  sender: {
    id: string;
    name: string;
    avatar: string | null;
  };
  content: string;
  attachement?: string | null;
  isMe?: boolean;
  createdAt: string;
};